# settings/backups/server/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Quester Multi-Tenant Authentication API
  version: 1.0.0
  description: |
    Secure JWT-based authentication system with multi-tenant support, RBAC, and gamification.
    
    ## Features
    - Multi-tenant architecture with tenant isolation
    - RS256 asymmetric JWT signing
    - Role-based access control (RBAC)
    - Token refresh with automatic rotation
    - Rate limiting on all endpoints
    - Gamification (XP, levels, tiers, login streaks)
    
    ## Authentication
    Most endpoints require a valid JWT access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
  contact:
    name: Quester API Support
    email: support@quester.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.quester.com/api/v1
    description: Production server
...
paths:
...
  /quests:
    get:
      tags:
        - Quests
      summary: List all active quests
      description: |
        Retrieves all quests for the current tenant.
        Supports filtering by status and type.
        
        **Rate Limit**: 100 requests per minute
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, expired]
        - name: type
          in: query
          schema:
            type: string
            enum: [course, achievement, challenge]
      responses:
        '200':
          description: List of quests
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      quests:
                        type: array
                        items:
                          $ref: '#/components/schemas/Quest'
                      tenant_id:
                        type: string
                        format: uuid
    
    post:
      tags:
        - Quests
      summary: Create a new quest (instructors/admins)
      description: |
        Creates a new quest in the system.
        
        **Rate Limit**: 20 requests per hour
        **Security**: Instructor or Admin role required
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - type
                - points
              properties:
                title:
                  type: string
                  example: "Complete JavaScript Basics"
                description:
                  type: string
                  example: "Finish all modules in JS course"
                type:
                  type: string
                  enum: [course, achievement, challenge]
                  example: "course"
                status:
                  type: string
                  enum: [active, draft, expired]
                  default: "draft"
                points:
                  type: integer
                  minimum: 0
                  example: 100
                  description: Gamification points awarded
                badge_id:
                  type: string
                  format: uuid
                  description: Optional badge awarded on completion
                requirements:
                  type: object
                  additionalProperties: true
                  description: Quest completion requirements
                start_date:
                  type: string
                  format: date-time
                end_date:
                  type: string
                  format: date-time
            example:
              title: "Complete JavaScript Basics"
              description: "Finish all modules in JS course"
              type: "course"
              points: 100
              requirements:
                course_id: "uuid-course-123"
                min_score: 80
      responses:
        '201':
          description: Quest created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Quest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /quests/{quest_id}:
    get:
      tags:
        - Quests
      summary: Get specific quest details
      description: |
        Retrieves detailed information about a quest, including progress for the current user.
        
        **Rate Limit**: 100 requests per minute
      security:
        - BearerAuth: []
      parameters:
        - name: quest_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Quest details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Quest'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - Quests
      summary: Update quest (instructors/admins)
      description: |
        Updates quest properties.
        
        **Rate Limit**: 20 requests per hour
        **Security**: Instructor or Admin role required
      security:
        - BearerAuth: []
      parameters:
        - name: quest_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [active, draft, expired]
                points:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: Quest updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Quest'
        '403':
          $ref: '#/components/responses/Forbidden'
    
    delete:
      tags:
        - Quests
      summary: Delete quest (moderators/admins)
      description: |
        Soft deletes a quest. Quest becomes inactive but data preserved.
        
        **Rate Limit**: 10 requests per hour
        **Security**: Moderator or Admin role required
      security:
        - BearerAuth: []
      parameters:
        - name: quest_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Quest deleted
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from signup, login, or refresh endpoints
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid email format
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
    Forbidden:
      description: Access forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: You do not have permission to access this resource
...